pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ctx"
version = "0.1.0"
description = "Filesystem-native context layer for AI agents — generates CONTEXT.md manifests"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "anthropic>=0.40",
    "openai>=1.50",
    "pyyaml>=6.0",
    "pathspec>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
ctx = "ctx.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
